% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presenceMatrix.R
\name{presenceMatrix}
\alias{presenceMatrix}
\title{Create a matrix of presences and absences}
\usage{
presenceMatrix(Data, Rows = "geoplate", Columns = "genus")
}
\arguments{
\item{Data}{A dataframe or matrix}

\item{Rows}{A characer string}

\item{Columns}{A character string}
}
\value{
A presence-absence matrix
}
\description{
Creates a community matrix of taxon presences and absences from a data frame with a column of sites and a column of species.
}
\examples{

# Download a test dataset of pleistocene bivalves.
DataPBDB<-downloadPBDB(Taxa="Bivalvia",StartInterval="Pleistocene",StopInterval="Pleistocene")

# Create a community matrix of genera by plates.
CommunityMatrix<-presenceMatrix(DataPBDB,Rows="geoplate",Columns="genus")

# Create a community matrix of families by geologic interval.
CommunityMatrix<-presenceMatrix(DataPBDB,Rows="early_interval",Columns="family")
}
\author{
Andrew A. Zaffos
}

